bubble plot integer attributes refactoring
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@2443 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/charts/xml/JRBubblePlotFactory.java
@@ -50,7 +50,7 @@ public class JRBubblePlotFactory extends JRBaseFactory {
 		String scaleType = attributes.getValue( ATTRIBUTE_scaleType );
 		Integer intScaleType = (Integer)JRXmlConstants.getScaleTypeMap().get( scaleType );
 		if( intScaleType != null ){
-			plot.setScaleType( intScaleType.intValue()  );
+			plot.setScaleType( intScaleType  );
 		}
 
 		return plot;
